기존 userId가 string 타입으로 정의된 부분 int로 변경

+++ src/main/java/com/aicodegem/repository/AchievementRepository.java
@@ -8,5 +8,5 @@ import java.util.List;
 // Achievement 엔티티에 대한 CRUD 및 사용자 정의 쿼리를 제공하는 리포지토리 인터페이스
 public interface AchievementRepository extends JpaRepository<Achievement, Long> {
     // 사용자 ID로 업적 조회
-    List<Achievement> findByUserId(int userId);
+    List<Achievement> findByUserId(Long userId);
 }
